+++ EDA/EDA.py
@@ -5,140 +5,9 @@ import numpy as np
 import diagnostics as diag
 import matplotlib.pyplot as plt
 import math
+import diagnostics as diag
+from on_off import find_time_course
 
-#functions
-def events2neural(task_fname, tr, n_trs):
-    """ Return predicted neural time course from event file `task_fname`
-
-    Parameters
-    ----------
-    task_fname : str
-        Filename of event file
-    tr : float
-        TR in seconds
-    n_trs : int
-        Number of TRs in functional run
-
-    Returns
-    -------
-    time_course : array shape (n_trs,)
-        Predicted neural time course, one value per TR
-    """
-    task = np.loadtxt(task_fname)
-    # Check that the file is plausibly a task file
-    if task.ndim != 2 or task.shape[1] != 3:
-        raise ValueError("Is {0} really a task file?", task_fname)
-    # Convert onset, duration seconds to TRs
-    task[:, :2] = task[:, :2] / tr
-    # Neural time course from onset, duration, amplitude for each event
-    time_course = np.zeros(n_trs)
-    for onset, duration, amplitude in task:
-        time_course[onset:math.ceil(onset + duration)] = amplitude
-    return time_course
-
-def vol_std(data):
-    """ Return standard deviation across voxels for 4D array `data`
-
-    Parameters
-    ----------
-    data : 4D array
-        4D array from FMRI run with last axis indexing volumes.  Call the shape
-        of this array (M, N, P, T) where T is the number of volumes.
-
-    Returns
-    -------
-    std_values : array shape (T,)
-        One dimensonal array where ``std_values[i]`` gives the standard
-        deviation of all voxels contained in ``data[..., i]``.
-    """
-
-    std_values=[]
-    for i in range(0,data.shape[-1]):
-        vol_1d = np.ravel(data[..., i])
-        std_values.append(np.std(vol_1d))
-    return std_values
-
-
-
-def iqr_outliers(arr_1d, iqr_scale=1.5):
-    """ Return indices of outliers identified by interquartile range
-
-    Parameters
-    ----------
-    arr_1d : 1D array
-        One-dimensional numpy array, from which we will identify outlier
-        values.
-    iqr_scale : float, optional
-        Scaling for IQR to set low and high thresholds.  Low threshold is given
-        by 25th centile value minus ``iqr_scale * IQR``, and high threshold id
-        given by 75 centile value plus ``iqr_scale * IQR``.
-
-    Returns
-    -------
-    outlier_indices : array
-        Array containing indices in `arr_1d` that contain outlier values.
-    lo_hi_thresh : tuple
-        Tuple containing 2 values (low threshold, high thresold) as described
-        above.
-    """
-    # Hint : np.lookfor('centile')
-    # Hint : np.lookfor('nonzero')
-
-    IQR = np.percentile(arr_1d,75) - np.percentile(arr_1d,25)
-    low = np.percentile(arr_1d,25)-iqr_scale*IQR
-    high = np.percentile(arr_1d,75)+iqr_scale*IQR
-    outlier_indices=np.where((arr_1d<low)|(arr_1d>high))[0]
-    lo_hi_thresh=(low,high)
-    return (outlier_indices,lo_hi_thresh)
-
-
-def vol_rms_diff(arr_4d):
-    """ Return root mean square of differences between sequential volumes
-
-    Parameters
-    ----------
-    data : 4D array
-        4D array from FMRI run with last axis indexing volumes.  Call the shape
-        of this array (M, N, P, T) where T is the number of volumes.
-
-    Returns
-    -------
-    rms_values : array shape (T-1,)
-        One dimensonal array where ``rms_values[i]`` gives the square root of
-        the mean (across voxels) of the squared difference between volume i and
-        volume i + 1.
-    """
-
-    rmsd_vals=[]
-    for i in range(0,arr_4d.shape[-1]-1):
-        diff_vol = arr_4d[..., i + 1] - arr_4d[..., i]
-        rmsd = np.sqrt(np.mean(diff_vol ** 2))
-        rmsd_vals.append(rmsd)
-    return rmsd_vals
-
-def extend_diff_outliers(diff_indices):
-    """ Extend difference-based outlier indices `diff_indices` by pairing
-
-    Parameters
-    ----------
-    diff_indices : array
-        Array of indices of differences that have been detected as outliers.  A
-        difference index of ``i`` refers to the difference between volume ``i``
-        and volume ``i + 1``.
-
-    Returns
-    -------
-    extended_indices : array
-        Array where each index ``j`` in `diff_indices has been replaced by two
-        indices, ``j`` and ``j+1``, unless ``j+1`` is present in
-        ``diff_indices``.  For example, if the input was ``[3, 7, 8, 12, 20]``,
-        ``[3, 4, 7, 8, 9, 12, 13, 20, 21]``.
-    """
-
-    extended_indices=[]
-    for i in diff_indices:
-        extended_indices.extend([i,i+1])
-    return np.unique(extended_indices)
 
 #loading data
 img = nib.load('bold.nii.gz')
@@ -149,14 +18,14 @@ data = data[..., 5:]
 ds2=data.shape
 print("1.The original data shape is %r. After dropping the first five, now the data has the shape %r") %(str(ds1),str(ds2))
 #standard deviations of all voxels along the TRs.
-std = vol_std(data)
+std = diag.vol_std(data)
 fobj = open('vol_std_values.txt', 'wt')
 for i in std:
 	fobj.write(str(i) + '\n')
 fobj.close()
 print("2. The standard deviations of all voxels along the TRs are saved in to 'vol_std_values.txt'")
 #find the std outliers
-outlier = iqr_outliers(std)[0]
+outlier = diag.iqr_outliers(std)[0]
 fobj = open('vol_std_outliers.txt', 'wt')
 for i in outlier:
 	fobj.write(str(i) + '\n')
@@ -165,8 +34,8 @@ print("3.There are %d std outliers, with indices %r. They are saved into 'vol_st
 
 #plot the std outliers
 std_outlier=[]
-low = iqr_outliers(std)[1][0]
-high = iqr_outliers(std)[1][1]
+low = diag.iqr_outliers(std)[1][0]
+high = diag.iqr_outliers(std)[1][1]
 for i in outlier:
 	std_outlier.append(std[i])
 x=np.arange(data.shape[-1])
@@ -182,13 +51,13 @@ plt.savefig('vol_std.png')
 plt.show()
 print("The std outliers is plotted and saved as 'vol_std.png'")
 #RMS diffrence
-rms = vol_rms_diff(data)
-rms_outlier = iqr_outliers(rms)[0]
+rms = diag.vol_rms_diff(data)
+rms_outlier = diag.iqr_outliers(rms)[0]
 rms_outlier_value = []
 for i in rms_outlier:
 	rms_outlier_value.append(rms[i])
-low_rms = iqr_outliers(rms)[1][0]
-high_rms = iqr_outliers(rms)[1][1]
+low_rms = diag.iqr_outliers(rms)[1][0]
+high_rms = diag.iqr_outliers(rms)[1][1]
 xx = np.arange(len(rms))
 plt.axis([0,140,0,25])
 rms1, = plt.plot(xx,rms,'b',label="rms values")
@@ -204,13 +73,13 @@ plt.show()
 print("4.There are %d RMS outliers with indices %r.") %(len(rms_outlier),rms_outlier)
 print("5.The RMS difference outliers is plotted and saved as 'vol_rms_outliers.png'")
 #extended RMS outliers
-ext_outlier = extend_diff_outliers(rms_outlier)
+ext_outlier = diag.extend_diff_outliers(rms_outlier)
 rms.append(0)
 ext_outlier_value = []
 for i in ext_outlier:
 	ext_outlier_value.append(rms[i])
-low_ext = iqr_outliers(rms)[1][0]
-high_ext = iqr_outliers(rms)[1][1]
+low_ext = diag.iqr_outliers(rms)[1][0]
+high_ext = diag.iqr_outliers(rms)[1][1]
 xxx = np.arange(len(rms))
 plt.axis([0,140,0,25])
 ext1, = plt.plot(xxx,rms,'b',label="rms values")
@@ -247,7 +116,7 @@ np.mean(data) #137.08845107920848
 #get the correlation matrix
 TR = 2.5
 n_trs = img.shape[-1]
-time_course = events2neural('cond002.txt', 2.5, n_trs) 
+time_course = find_time_course('cond002.txt', 2.5, n_trs) 
 plt.plot(time_course)
 plt.title("time_course")
 plt.savefig("time_course.png")
