+++ EDA/on_off.py
@@ -1,0 +1,31 @@
+import numpy as np
+import math
+
+def find_time_course (task_fname, tr, n_trs):
+    """ Return predicted neural time course from event file `task_fname`
+
+    Parameters
+    ----------
+    task_fname : str
+        Filename of event file
+    tr : float
+        TR in seconds
+    n_trs : int
+        Number of TRs in functional run
+
+    Returns
+    -------
+    time_course : array shape (n_trs,)
+        Predicted neural time course, one value per TR
+    """
+    task = np.loadtxt(task_fname)
+    # Check that the file is plausibly a task file
+    if task.ndim != 2 or task.shape[1] != 3:
+        raise ValueError("Is {0} really a task file?", task_fname)
+    # Convert onset, duration seconds to TRs
+    task[:, :2] = task[:, :2] / tr
+    # Neural time course from onset, duration, amplitude for each event
+    time_course = np.zeros(n_trs)
+    for onset, duration, amplitude in task:
+        time_course[onset:math.ceil(onset + duration)] = amplitude
+    return time_course
